refactor: centralize popup modals
