+++ src/components/FinancialSummary/FinancialOverviewCards/BankBalanceCard.jsx
@@ -4,7 +4,7 @@ import React, { useContext } from 'react';
 import { Card, Col, Space, Statistic, Typography, Tooltip } from 'antd';
 import { IconBuildingBank } from '@tabler/icons-react';
 import { FinanceContext } from '../../../contexts/FinanceContext';
-import BankBalanceEditModal from './BankBalanceEditModal';
+import BankBalanceEditModal from '../../PopUpModals/BankBalanceEditModal';
 import './BankBalanceCard.css';
 
 const { Text } = Typography;
